updated outputs - still needs testing

+++ EmailSqlResults/OutputQuery.cs
@@ -6,24 +6,48 @@ using System.Text;
 using System.Threading.Tasks;
 using excel = Microsoft.Office.Interop.Excel;
 using System.IO;
-using System.Data;
 using System.Windows.Forms;
 using System.Runtime.InteropServices;
 
 namespace EmailSqlResults
 {
-    class ExcelPush
+    class PushData
     {
         public bool isSuccessful { get; set; }
+        public System.Data.DataTable dT { get; set; }
+        public string location { get; set; }
 
-        public ExcelPush(System.Data.DataTable Dt, string Location)
+        public PushData(System.Data.DataTable Dt, string Location)
         {
-            if (File.Exists(Location))
-                File.Delete(Location);
-            WriteToExcel(Dt, Location);
+            this.dT = Dt;
+            this.location = Location;
+            if (File.Exists(Location)) { File.Delete(Location); }
         }
 
-        private void WriteToExcel(System.Data.DataTable dt, string location)
+        public void WriteToCsv(string sepString)        
+        {
+            //exports Datatable as csv
+            var sb = new StringBuilder();
+            // create columns
+            for (int i = 0; i < dT.Columns.Count; i++)
+            {
+                sb.Append(dT.Columns[i].ColumnName);
+                sb.Append(i == dT.Columns.Count - 1 ? "" : sepString);
+            }            
+            sb.Append("\n");
+            for (int currentRow = 0; currentRow < dT.Rows.Count; currentRow++)
+            {
+                for (int j = 0; j < dT.Columns.Count; j++)
+                {
+                    sb.Append(dT.Rows[currentRow][j].ToString());
+                    sb.Append(j == dT.Columns.Count - 1 ? "" : sepString);
+                }
+                if (currentRow < dT.Rows.Count - 1) { sb.Append("\n"); }
+            }
+            File.AppendAllText(location + ".csv", sb.ToString());
+        }
+
+        public void WriteToExcel()
         {
             //instantiate excel objects (application, workbook, worksheets)
             excel.Application XlObj = new excel.Application();
@@ -35,7 +59,7 @@ namespace EmailSqlResults
             try
             {
                 int row = 1; int col = 1;
-                foreach (DataColumn column in dt.Columns)
+                foreach (System.Data.DataColumn column in dT.Columns)
                 {
                     //adding columns
                     WsObj.Cells[row, col] = column.ColumnName;
@@ -44,10 +68,10 @@ namespace EmailSqlResults
                 //reset column and row variables
                 col = 1;
                 row++;
-                for (int i = 0; i < dt.Rows.Count; i++)
+                for (int i = 0; i < dT.Rows.Count; i++)
                 {
                     //adding data
-                    foreach (var cell in dt.Rows[i].ItemArray)
+                    foreach (var cell in dT.Rows[i].ItemArray)
                     {
                         WsObj.Cells[row, col] = cell;
                         col++;
@@ -55,7 +79,7 @@ namespace EmailSqlResults
                     col = 1;
                     row++;
                 }
-                WbObj.SaveAs(location);
+                WbObj.SaveAs(location + ".xlsx");
                 isSuccessful = true;
             }
             catch (COMException cEx)
@@ -74,6 +98,12 @@ namespace EmailSqlResults
                 XlObj.Quit();                
             }
         }
+
+
+
+
     }
+
+
 }
